Delegate and extend instead of copying lib code.
Gahhhh the one time you want multiple inheritance.
Might actually be nice to switch it so that the IntegrationBase extends FluentTestNg, 
since it really "is a" test and could "have a" TestNGSpringContextTests.
That said it truly is both of those things, but the benefit of making it delegate to TestNGSpringContextTests is that you'd have far fewer methods to delegate.

This solution, while better than copy-pasting the AbstractTestNGSpringContextTests, is still kinda gross.